requirements.txt
Django==2.2.12
djangorestframework==3.12.4
requests==2.22.0
simplejson==3.16.0
gunicorn==20.1.0